--- homeworks/week22/hw1/src/components/Header/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Header";
