+++ homeworks/week22/hw1/src/components/LoginForm/LoginForm.js
@@ -8,6 +8,7 @@ const Container = styled.div`
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
+  margin-top: 80px;
 `;
 
 const Title = styled.div`
