+++ homeworks/week22/hw1/src/context.js
@@ -2,3 +2,4 @@ import { createContext } from "react";
 
 export const AuthContext = createContext(null);
 export const LoadingContext = createContext(null);
+export const PageContext = createContext(null);
