+++ homeworks/week22/hw1/src/pages/HomePage/HomePage.js
@@ -1,8 +1,10 @@
-import { useState, useEffect, useCallback } from "react";
+import { useContext } from "react";
 import styled from "styled-components";
-import { getPosts, countTotalPosts } from "../../WebAPI";
 import Post from "../../components/Post";
 import Paginator from "../../components/Paginator";
+import useGetPost from "../../customHooks/useGetPosts";
+import usePaginate from "../../customHooks/usePaginate";
+import { PageContext } from "../../context";
 
 const Root = styled.div``;
 
@@ -13,49 +15,15 @@ const PostList = styled.div`
   overflow: scroll;
 
   @media screen and (max-width: 768px) {
-    margin: 100px auto;
+    margin: 100px auto 60px auto;
     width: 80%;
   }
 `;
 
 function HomePage() {
-  const [posts, setPosts] = useState([]);
-  const [page, setPage] = useState(1);
-  const [totalPage, setTotalPage] = useState();
-
-  useEffect(() => {
-    countTotalPosts().then((posts) => {
-      const totalPost = Number(posts.length);
-      setTotalPage(Math.ceil(totalPost / 5));
-    });
-  }, [posts]);
-
-  useEffect(() => {
-    getPosts(page).then((posts) => {
-      setPosts(posts);
-    });
-  }, [page]);
-
-  const handleButtonClick = useCallback(
-    (button) => {
-      if (button === "First") {
-        return setPage(1);
-      }
-
-      if (button === "Prev" && page > 1) {
-        return setPage(page - 1);
-      }
-
-      if (button === "Next" && page < totalPage) {
-        return setPage(page + 1);
-      }
-
-      if (button === "Last") {
-        return setPage(totalPage);
-      }
-    },
-    [page, totalPage]
-  );
+  const { posts } = useGetPost();
+  const { handleButtonClick } = usePaginate();
+  const { page } = useContext(PageContext);
 
   return (
     <Root>
